+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-02-21
+### Added
+- New function `create-request-method-handler`
+### Changed
+- Renamed `namespaced-handler-fn` to `method-handler-fn`
+
 ## 0.1.0 - 2019-02-14
 ### Added
 - Initial public release
 
-[Unreleased]: https://github.com/sbreatnach/treo/compare/0.1.0...HEAD
+[0.2.0]: https://github.com/sbreatnach/treo/compare/0.1.0...0.2.0
+[Unreleased]: https://github.com/sbreatnach/treo/compare/0.2.0...HEAD